new video bootstrap and embed entry point
