+++ applications/app/views/videoEmbed.scala.html
@@ -100,8 +100,17 @@
         document.documentElement.className = docClass;
         </script>
 
-        @if(VideoJSSwitch.isSwitchedOn) {
-            <script>
+    @if(VideoJSSwitch.isSwitchedOn) {
+        <script>
+            (function (document) {
+                var script = document.createElement('script');
+                var ref = document.getElementsByTagName('script')[0];
+                script.src = '@Static("javascripts/graun.videojs-embed.js")';
+                ref.parentNode.insertBefore(script, ref);
+            })(document);
+            </script>
+    } else {
+        <script>
             (function (document) {
                 var script = document.createElement('script');
                 var ref = document.getElementsByTagName('script')[0];
@@ -109,7 +118,7 @@
                 ref.parentNode.insertBefore(script, ref);
             })(document);
             </script>
-        }
+    }
 
         @fragments.analytics.base()(page, request, context)
