+++ static/src/javascripts/bootstraps/enhanced/video-player.js
@@ -1,0 +1,3 @@
+export const initVideoPlayer = (): void => {
+    console.log('Enhancing video');
+};
